update the run about 100 structures
With the update of script smiles_obsdf.py, the analysis with the
set of 100 molecules was repeated.  The same number of entries,
indeed the same structures as with the prevous version of the
script pass/do not pass the round trip edit (criterion: the same
canonical SMILES assigned by OpenBabel prior and after the round
trip edit).