extend the round trip to a set of 5k molecules
With a library of 5k drug-like random molecules, the three types of
a round trip are performed.  With the criterion of an invariant
InChI string, the rate of success remains lower, than anticipated
(41.0% smiles_rdkit.py, 40.5% smiles_obabel.py, 36.1% inchi_sdf.py)
indicating a systematic problem how the round trip is implemented.