update processing w/ Jmol and awk
The probe is Jmol's assigned SMILES prior and after the round trip.

Testing criteria previously tested vs. the set of 100 molecules
e.g., definitions based on check of primary/secondary InChI, are
only muted.  Though Robert Hanson contributed by constructive
criticsm, a separate project about an InChI round trip with Jmol
might be an advantage for further work on the Jmol script.

AWK is used to split the output in separate lists, the one about
"failing" in a format suitable for visual inspection in CDKdepict.

It is a cherry-pick from branch Jmol2, commitsha fe3e502.